fix: 테스트 분류에 따른 패키지명 구성
