+++ src/test/java/team1/BE/seamless/e2e/UserTokenParseTest.java
@@ -1,13 +1,8 @@
-package team1.BE.seamless.service;
+package team1.BE.seamless.e2e;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.springframework.http.HttpMethod.DELETE;
 import static org.springframework.http.HttpMethod.POST;
-import static org.springframework.http.HttpMethod.PUT;
-import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
@@ -16,10 +11,9 @@ import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import team1.BE.seamless.DTO.UserDTO.UserUpdate;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class ProjectServiceTest {
+class UserTokenParseTest {
 
     @LocalServerPort
     private int port;
@@ -28,13 +22,10 @@ class ProjectServiceTest {
     private String token;
     private HttpHeaders headers = new HttpHeaders();
 
-    private final ProjectService projectService;
-
 
     @Autowired
-    public ProjectServiceTest(TestRestTemplate restTemplate, ProjectService projectService) {
+    public UserTokenParseTest(TestRestTemplate restTemplate) {
         this.restTemplate = restTemplate;
-        this.projectService = projectService;
     }
 
     /**
@@ -55,4 +46,5 @@ class ProjectServiceTest {
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setBearerAuth(token);
     }
+
 }
